crowdfunding/requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==3.2.16
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
dotenv==0.9.9
gunicorn==20.1.0
iniconfig==2.3.0
packaging==25.0
pillow==10.2.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.5
